Python
1. We should be able to run the Python script from the command line and it should
open a Web Page in the browser with a Descent and Usable Interface.
2. User should have an option on the Web Page to close the opened browser after his
work is complete, it should also close/kill the script running in the background.
3. UI should have a Map tool (Interactive Visualisation) to generate a Scatter Geo
Map with markers for highlighting the lat/long where the incident happened based
on combinations of Month, Day, Attack Type, Region, Country, State and City.
User should be able to select any DAY and MONTH and based on that DAY and
MONTH we need to show what incidents happened on that DAY and MONTH (
maybe in any year ). We need to create a single visualization of every combination
possible and it should get updated accordingly.
4. Client should be able to filter the 9 Type of Attacks ( Bombing, Assassination,
Kidnapping etc )
5. Client should be able to filter the Region, Country, State, City.
6. Hovering and Clicking of the mouse should have information.
7. Separate for World and India
8. Option to have animation to Play and Pause for the Years and also should be able
to filter data based on year range slider.

9. UI should have a Chart tool (Interactive Visualisation) to show Stacked Line
Chart images of the frequency of terrorist incidents each year.
10. Client should be able to Group first by (Country Attacked, Region, Target
Nationality, Target Type, Type of Attack, Weapon Type, Terrorist Organisation)
11. Client should be able to search based on the selected Group
12. Client should be able to Filter data based on the Year
13. Client should be able to see the detailing in a visualisation.
14. Follow Rule (overview first, zoom, Compare and filter, details on demand )
15. When a user's mouse hovers over a stripe we should show details
16. Separate for World and India
17. Option to have animation to Play and Pause for the Years and also should be
able to filter data based on year range slider.